requirements.txt
python-telegram-bot==13.15
PyPDF2==3.0.1
deep-translator==1.8.1
